requirements.txt
paho-mqtt
pandas
protobuf
googleapis-common-protos
